+++ crates/plotinator-file-io/src/file_dialog.rs
@@ -5,5 +5,6 @@ pub mod web;
 
 /// Used for file dialog filters
 pub const FILE_FILTER_NAME: &str = "Known Logs and Plotinator3000 files";
-pub const FILE_FILTER_EXTENSIONS: &[&str] =
-    &["p3k", "sps", "h5", "hdf5", "hdf", "bin", "log", "zip"];
+pub const FILE_FILTER_EXTENSIONS: &[&str] = &[
+    "p3k", "sps", "h5", "hdf5", "hdf", "bin", "log", "zip", "csv", "txt",
+];